# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Təhvil renovation project coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/change-orders:
    get:
      operationId: listChangeOrders
      tags: [changes]
      summary: List project change orders
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Change orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeOrder"
    post:
      operationId: createChangeOrder
      tags: [changes]
      summary: Create and submit a change order
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeOrderInput"
      responses:
        "201":
          description: Change order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeOrder"
  /change-orders/{changeOrderId}/decision:
    post:
      operationId: decideChangeOrder
      tags: [changes]
      summary: Approve, reject, or request clarification on a change order
      parameters:
        - name: changeOrderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeOrderDecision"
      responses:
        "200":
          description: Change order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeOrder"
...
  /projects/{projectId}/timeline:
    get:
      operationId: listTimeline
      tags: [activity]
      summary: Get the append-only audit timeline
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /projects/{projectId}/invitations:
    post:
      operationId: createInvitation
      tags: [activity]
      summary: Invite a contractor or viewer to a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitation created with a one-time share URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [activity]
      summary: Accept a valid project invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationAcceptInput"
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}/passport:
    get:
      operationId: getRenovationPassport
      tags: [activity]
      summary: Get the printable renovation passport summary
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project archive summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passport"
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        propertyType: { type: string }
        city: { type: string }
        projectType: { type: string }
        plannedStartDate: { type: ["string", "null"], format: date }
        plannedCompletionDate: { type: ["string", "null"], format: date }
        budget: { type: ["number", "null"] }
        approvedBudget: { type: number }
        approvedChanges: { type: number }
        progress: { type: integer, minimum: 0, maximum: 100 }
        status: { type: string }
        scopeVersion: { type: integer }
        scopeStatus: { type: string, enum: [draft, pending_approval, approved] }
        archived: { type: boolean }
        participantRole: { type: string, enum: [owner, contractor, viewer] }
        contractorName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, propertyType, city, projectType, approvedBudget, approvedChanges, progress, status, scopeVersion, scopeStatus, archived, participantRole, createdAt]
...
    Room:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        itemCount: { type: integer }
      required: [id, projectId, name, itemCount]
...
    ScopeItem:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        roomId: { type: ["string", "null"] }
        roomName: { type: ["string", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        inclusionType: { type: string, enum: [included, excluded] }
        materialResponsibility: { type: string, enum: [owner, contractor, shared, not_applicable] }
        laborAmount: { type: ["number", "null"] }
        materialEstimate: { type: ["number", "null"] }
        warrantyMonths: { type: ["integer", "null"] }
        status: { type: string, enum: [draft, active, completed, excluded] }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, title, inclusionType, materialResponsibility, status, createdAt]
...
    ChangeOrder:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        number: { type: integer }
        roomName: { type: ["string", "null"] }
        title: { type: string }
        reason: { type: string }
        description: { type: ["string", "null"] }
        laborAmount: { type: number }
        materialAmount: { type: number }
        totalAmount: { type: number }
        scheduleImpactDays: { type: integer }
        newCompletionDate: { type: ["string", "null"], format: date }
        status: { type: string, enum: [submitted, needs_clarification, approved, rejected] }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        decisionComment: { type: ["string", "null"] }
      required: [id, projectId, number, title, reason, laborAmount, materialAmount, totalAmount, scheduleImpactDays, status, createdBy, createdAt]
    ChangeOrderInput:
      type: object
      properties:
        roomId: { type: ["string", "null"] }
        title: { type: string, minLength: 2, maxLength: 160 }
        reason: { type: string, minLength: 2, maxLength: 1000 }
        description: { type: ["string", "null"], maxLength: 3000 }
        laborAmount: { type: number, minimum: 0 }
        materialAmount: { type: number, minimum: 0 }
        scheduleImpactDays: { type: integer, minimum: 0 }
        newCompletionDate: { type: ["string", "null"], format: date }
      required: [title, reason, laborAmount, materialAmount, scheduleImpactDays]
    ChangeOrderDecision:
      type: object
      properties:
        decision: { type: string, enum: [approve, reject, clarify] }
        comment: { type: ["string", "null"], maxLength: 1000 }
      required: [decision]
    Milestone:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        plannedCompletionDate: { type: ["string", "null"], format: date }
        status: { type: string, enum: [planned, in_progress, submitted_for_handover, revision_requested, accepted] }
        note: { type: ["string", "null"] }
        media:
          type: array
          items: { $ref: "#/components/schemas/MediaItem" }
        acceptedAt: { type: ["string", "null"], format: date-time }
      required: [id, projectId, title, status, media]
...
    MediaItem:
      type: object
      properties:
        id: { type: string }
        objectPath: { type: string }
        originalName: { type: string }
        contentType: { type: string }
        size: { type: integer }
        uploadedAt: { type: string, format: date-time }
      required: [id, objectPath, originalName, contentType, size, uploadedAt]
...
    PaymentItem:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        amount: { type: number }
        currency: { type: string, enum: [AZN] }
        dueCondition: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"], format: date }
        status: { type: string, enum: [planned, due, marked_sent, confirmed_received, disputed] }
        note: { type: ["string", "null"] }
        media:
          type: array
          items: { $ref: "#/components/schemas/MediaItem" }
        updatedAt: { type: string, format: date-time }
      required: [id, projectId, title, amount, currency, status, media, updatedAt]
...
    TimelineEvent:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        action: { type: string }
        detail: { type: string }
        actorRole: { type: string, enum: [owner, contractor, viewer] }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, action, detail, actorRole, createdAt]
    Invitation:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        inviteUrl: { type: string }
        role: { type: string, enum: [contractor, viewer] }
        expiresAt: { type: string, format: date-time }
      required: [id, projectId, inviteUrl, role, expiresAt]
    InvitationInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        contact: { type: string, minLength: 3, maxLength: 254 }
        role: { type: string, enum: [contractor, viewer] }
      required: [name, contact, role]
    InvitationAcceptInput:
      type: object
      properties:
        token: { type: string, minLength: 20 }
      required: [token]
...
    Passport:
      type: object
      properties:
        project: { $ref: "#/components/schemas/Project" }
        rooms:
          type: array
          items: { $ref: "#/components/schemas/Room" }
        scopeItems:
          type: array
          items: { $ref: "#/components/schemas/ScopeItem" }
        changeOrders:
          type: array
          items: { $ref: "#/components/schemas/ChangeOrder" }
        milestones:
          type: array
          items: { $ref: "#/components/schemas/Milestone" }
        payments:
          type: array
          items: { $ref: "#/components/schemas/PaymentItem" }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/TimelineEvent" }
        disclaimer: { type: string }
      required: [project, rooms, scopeItems, changeOrders, milestones, payments, timeline, disclaimer]
...