# lib/api-zod/src/generated/types/changeOrderDecision.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Təhvil renovation project coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeOrderDecisionDecision } from './changeOrderDecisionDecision';

export interface ChangeOrderDecision {
  decision: ChangeOrderDecisionDecision;
  /**
     * @maxLength 1000
     * @nullable
     */
  comment?: string | null;
}